refactor: :art: Improve structure of the project
